script files simplification

+++ scripts/aktualizuj.bash
@@ -7,7 +7,13 @@ then
 	exit
 fi
 
-wget https://raw.githubusercontent.com/RCPRG-ros-pkg/irp6_robot/master/scripts/irp6.rosinstall -O /tmp/irp6.rosinstall
+if test -e ./hardware; then
+  wget https://raw.githubusercontent.com/RCPRG-ros-pkg/irp6_robot/master/scripts/irp6.rosinstall -O /tmp/irp6.rosinstall
+else
+ wget https://raw.githubusercontent.com/RCPRG-ros-pkg/irp6_robot/master/scripts/irp6_sim.rosinstall -O /tmp/irp6.rosinstall
+fi
+
+
 wget https://raw.githubusercontent.com/RCPRG-ros-pkg/irp6_robot/master/scripts/update_and_compile.bash -O /tmp/update_and_compile.bash
